% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bliss}
\alias{bliss}
\title{"Bliss, and Other Stories"}
\format{A character vector with 1622 elements}
\source{
\ulr{https://ebooks.adelaide.edu.au/m/mansfield/katherine/bliss/index.html}
}
\usage{
bliss
}
\description{
A dataset containing the full text of "Bliss, 
and Other Stories" by Katherine Mansfield. The UTF-8 plain text was
sourced from the ebooks.adelaide.edu website. Each line represents a 
paragraph in the original text.
}
\keyword{datasets}

