% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{charfreq}
\alias{charfreq}
\title{Count frequency of word or punctuation mark in a text}
\usage{
charfreq(characters, char.list, punctuation = FALSE)
}
\arguments{
\item{characters}{Large vector of words/punctuation marks to be matched with 
a smaller list of desired words/characters.}

\item{char.list}{Vector of target words/punctuation marks to be matched with the larger
vector of words/characters}

\item{punctuation}{Boolean that determines whether to convert
punctuation marks into words}
}
\description{
Matches a vector of target words/punctuation marks to a larger vector 
of words/punctuation marks and counts how many times that particular word or 
punctuation marks occurs in the larger string. Returns a data table with the 
matched words or punctuation marks with their number the occurrances 
in the larger string.
}
\note{
the accepted punctuation marks are commas, periods, semicolons, question marks
       exclamation points, quotation marks (forward and backward), ellipses and em-dashes.
}
\examples{
char <- extract_token(gardenParty)
charfreq(char, c("she", "he", "them"), punctuation = FALSE)
     character freq
[1]       she  1349
[2]        he  8632
[3]      them  223

char <- extract_punct(gardenParty)
charfreq(char, c(".", "...", "?"), punctuation = TRUE)
          character    freq
[1]          period   15229
[2]        ellipsis     199
[3]   question_mark     603
}

