% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_prep.R
\name{collapse_text}
\alias{collapse_text}
\title{Collapse text}
\usage{
collapse_text(text)
}
\arguments{
\item{text}{Character vector containing all the lines in a given text}
}
\description{
This function retrieves a vector of character string objects 
and collapses it into one string.
}
\examples{
collapse(text = "gardenParty")
}

