% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{completeAnalysisTable}
\alias{completeAnalysisTable}
\title{Analysis Table of Katheirne Mansfield's Complete Works}
\format{A data table with 54 observations (each row representing one story) and
27 variables (columns)}
\usage{
completeAnalysisTable
}
\description{
A dataset containing the make_analysis_df() output for 
the mansfieldComplete dataset, including a year and period variable.
}
\keyword{datasets}

