% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_punct}
\alias{extract_punct}
\title{extract punctuation}
\usage{
extract_punct(text)
}
\arguments{
\item{text}{character vector}

\item{breaks}{vector of strings}
}
\description{
given a string and a list of punc
}
\examples{
extract_sentence(text=gardenParty)
}

