% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_sentence}
\alias{extract_sentence}
\title{Extract sentences}
\usage{
extract_sentence(text)
}
\arguments{
\item{text}{Random string}
}
\description{
Takes a string and returns a vector
of each sentence in the string.
}
\examples{
extract_sentence(text="gardenParty")
}

