% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_sentences}
\alias{extract_sentences}
\title{Extract sentences}
\usage{
extract_sentences(text)
}
\arguments{
\item{text}{A random string, as a character vector}
}
\description{
Takes a string and returns a vector
of each sentence in the string.
}
\note{
This function takes out all periods, question marks, 
      exclamation points and quotation marks.
}
\examples{
sampleText <- gardenParty[10]
extract_sentences(text = sampleText)
}

