% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_token}
\alias{extract_token}
\title{Extract tokens}
\usage{
extract_token(text)
}
\arguments{
\item{text}{Random string}
}
\description{
Takes a string and returns a vector
of every word in the string.
}
\examples{
extract_token(text=gardenParty)
}

