% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_type}
\alias{extract_type}
\title{Extract types}
\usage{
extract_type(text)
}
\arguments{
\item{text}{Random string}
}
\description{
Takes a string and returns a vector
of each unique word in the string.
}
\examples{
extract_type(text=gardenParty)
}

