% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_prep.R
\name{find_chapters}
\alias{find_chapters}
\title{Find chapter breaks in Mansfield data}
\usage{
find_chapters(text)
}
\arguments{
\item{text}{Character vector containing all the lines in a given text}
}
\description{
Returns the line index numbers of the first lines of each short story in
the vectorized short story collections. Only works for the Katherine Mansfield
data included in the katherinemansfieldr package.
}
\note{
The last element in the output is the line
}
\examples{
find_chapters(text = gardenParty)
}

