% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{freq_punct_line}
\alias{freq_punct_line}
\title{Get frequency of punctuation marks per line}
\usage{
freq_punct_line(text, punctlist)
}
\arguments{
\item{text}{Vector of strings representing lines of a text}

\item{punctlist}{Vector of target punctuation marks to be matched 
with the text/document of interest}
}
\description{
Returns a data table returning the number of times a series of 
desired punctuation marks appear in a given line and the index 
number of that line.
}
\examples{
freq_punct_line(text = gardenParty, punctlist = c(".", "?", "..."))
}

