% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{freqPunct_line}
\alias{freqPunct_line}
\title{Gets frequency of punctuation per line}
\usage{
freqPunct_line(text, punctlist)
}
\arguments{
\item{text}{Vector of strings representing lines of a text}

\item{punctlist}{Punctuation marks whose frequency we are 
interested in finding}
}
\description{
Returns a data table returning the frequency at which a series of 
desired punctuation marks appear in a given line and the index 
number of that line.
}
\examples{
freqPunct_line(text = gardenParty, punctlist = c(".", "?", "..."))
}

