% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{freqWord_line}
\alias{freqWord_line}
\title{Gets frequency of words per line}
\usage{
freqWord_line(text, freqwords)
}
\arguments{
\item{text}{Vector of strings representing lines of a text}

\item{freqwords}{Words whose frequency we are interested in finding}
}
\description{
Returns a data table returning the frequency at which a series of desired
words appear in a given line and the index number of that line.
}
\examples{
freqWord_line(text = gardenParty, freqwords = c("she", "they", "he"))
}

