% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gardenParty}
\alias{gardenParty}
\title{"The Garden Party, and Other Stories"}
\format{A character vector with 1541 elements}
\source{
\url{https://ebooks.adelaide.edu.au/m/mansfield/katherine/garden/index.html}
}
\usage{
gardenParty
}
\description{
A dataset containing the full text of "The Garden Party, 
and Other Stories" by Katherine Mansfield. The UTF-8 plain text was
sourced from the ebooks.adelaide.edu.au website. Each line represents a 
paragraph in the original text.
}
\keyword{datasets}

