% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_prep.R
\name{get_breaks}
\alias{get_breaks}
\title{Break text by line matching}
\usage{
get_breaks(text, breaks)
}
\arguments{
\item{text}{Character vector containing all the lines in a given text}

\item{breaks}{Lines to be matched in the text}
}
\description{
This function matches a vector of break point strings 
with lines from another text and returns 
the line/index numbers of the matched strings in the text.
}
\examples{
get_breaks(text="gardenParty",breaks="breakpoints")
}

