% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mansfieldComplete}
\alias{mansfieldComplete}
\title{Katherine Mansfield's Complete Works}
\format{A character vector with 4574 elements}
\source{
\url{https://ebooks.adelaide.edu.au/m/mansfield/katherine/}
}
\usage{
mansfieldComplete
}
\description{
A dataset containing the full texts of "Something Childish, 
and Other Stories", "Bliss, and Other Stories", and "The Garden
Party, and Other Stories" by Katherine Mansfield. The UTF-8 plain text was
sourced from the ebooks.adelaide.edu.au website. Each line represents a 
paragraph in the original text.
}
\keyword{datasets}

