% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_prep.R
\name{multi_scan}
\alias{multi_scan}
\title{Scan multiple text files into character vector}
\usage{
multi_scan(list, directory)
}
\arguments{
\item{directory}{Local directory from which text files will be scanned}

\item{text}{Character vector containing all the lines in a given text}
}
\description{
This function retrieves the names of a series of text files and
returns a vector composed of the lines of all the listed 
files.
}
\examples{
find_freq_char(list=c("chapter1.txt", "chapter2.txt"), directory="~/STAT495-Group6/textFiles")
}

