% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_length.R
\name{sentence_length}
\alias{sentence_length}
\title{Get sentence length}
\usage{
sentence_length(sentences)
}
\arguments{
\item{sentences}{A character vector of sentences}
}
\description{
Retrieves a vector of sentences and returns the length of 
each sentence in terms of number of words.
}
\examples{
sentence_length(sentences=c("And after all the weather was ideal.", 
"Windless, warm, the sky without a cloud."))
[1] 7
[2] 7
}

