% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_length.R
\name{sentence_length_line}
\alias{sentence_length_line}
\title{Gets sentence length per line}
\usage{
sentence_length_line(text)
}
\arguments{
\item{text}{Vector of strings representing lines of a text}
}
\description{
Returns a data table returning the average sentence length in a line
in terms of number of words and the index number of that line.
}
\examples{
sentence_length_line(text=gardenParty)
}

