% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{somethingChildish}
\alias{somethingChildish}
\title{"Something Childish, and Other Stories"}
\format{A character vector with 1411 elements}
\source{
\url{https://ebooks.adelaide.edu.au/m/mansfield/katherine/something/index.html}
}
\usage{
somethingChildish
}
\description{
A dataset containing the full text of "Something Childish, 
and Other Stories" by Katherine Mansfield. The UTF-8 plain text was
sourced from the ebooks.adelaide.edu website. Each line represents a 
paragraph in the original text.
}
\keyword{datasets}

