% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_length.R
\name{word_length}
\alias{word_length}
\title{Get token length}
\usage{
word_length(words)
}
\arguments{
\item{words}{A vector of words}
}
\description{
Retrieves a vector of tokens (words) and returns a vector 
containing the length of each token in terms of number of
unique characters.
}
\examples{
word_length(words=c("very", "early", "morning"))
}

