% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_length.R
\name{word_length_line}
\alias{word_length_line}
\title{Get token length per line}
\usage{
word_length_line(text)
}
\arguments{
\item{text}{Vector of strings, with each element representing one line from a text}
}
\description{
Returns a vector containing the average token (word) length in a line of text
in terms of number of characters.
}
\examples{
word_length_line(text = gardenParty)
}

