% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{charfreq}
\alias{charfreq}
\title{Counts frequency of character or word in a text}
\usage{
charfreq(characters, char.list, punctuation = FALSE)
}
\arguments{
\item{characters}{Words to be matched with a smaller list of desired 
words}

\item{char.list}{Words whose frequency we are interested in measuring}

\item{punctuation}{Boolean that determines whether to convert
punctuation marks into words}
}
\description{
This function matches a list of strings to a text and counts how 
many times a particular string occurs in the text. Returns a data
table with the matched strings with its number the occurrances.
}
\examples{
sentence_length(sentences=c("And after all the weather was ideal.", 
"Windless, warm, the sky without a cloud."))
}

