% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_prep.R
\name{collapse_text}
\alias{collapse_text}
\title{Collapse text}
\usage{
collapse_text(text)
}
\arguments{
\item{text}{Character vector containing all the lines in a given text}
}
\description{
Retrieves a character vector collapses it into one string.
}
\examples{
sampleText <- c("“Oh, Mrs. Parker, I’m going out.”", "“Very good, sir.”")
collapse_text(text = sampleText)
[1] "“Oh, Mrs. Parker, I’m going out.” “Very good, sir.”"
}

