% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_punct}
\alias{extract_punct}
\title{extract punctuation}
\usage{
extract_punct(text)
}
\arguments{
\item{text}{Character vector containing all the lines in a given text}
}
\description{
Takes a string and returns a vector of every
punctuation mark present in the string. The punctuation
marks returned are: periods, commas, semicolons, question 
marks, exclamation points, quotation marks (both forward
and backward), ellipses and em-dashes.
}
\examples{
sampleText <- gardenParty[10]
extract_punct(text = sampleText)
}

