% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_token}
\alias{extract_token}
\title{Extract tokens}
\usage{
extract_token(text)
}
\arguments{
\item{text}{A random string, as a character vector}
}
\description{
Takes a string and returns a vector
with every word in the string as a separate
element.
}
\examples{
sampleText <- gardenParty[10]
extract_token(text = sampleText)
}

