% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_type}
\alias{extract_type}
\title{Extract types}
\usage{
extract_type(text)
}
\arguments{
\item{text}{Character vector containing all the lines in a given text}
}
\description{
Takes a string and returns a vector
with every unique word in the string as
a separate element.
}
\examples{
sampleText <- gardenParty[10]
extract_type(text = sampleText)
}

