% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_prep.R
\name{find_freq_char}
\alias{find_freq_char}
\title{Finds 10 most frequent words}
\usage{
find_freq_char(text)
}
\arguments{
\item{text}{Character vector containing all the lines in a given text}
}
\description{
This function retrieves the 10 most frequently occurring
words and returns the words in a vector.
}
\examples{
find_freq_char(text = "gardenParty")
}

