% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{freq_word_line}
\alias{freq_word_line}
\title{Get frequency of words per line}
\usage{
freq_word_line(text, freqwords)
}
\arguments{
\item{text}{Any text or document as a character vector}

\item{freqwords}{Vector of target words 
to be matched with the text/document of interest}
}
\description{
Returns a data table returning the number of times a series of desired
words appear in a given line of text and the index number of that line.
}
\examples{
freq_word_line(text = gardenParty, freqwords = c("she", "they", "he"))
}

