% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variety.R
\name{get_variety}
\alias{get_variety}
\title{Gets variety value}
\usage{
get_variety(text)
}
\arguments{
\item{text}{String or vector of strings}
}
\description{
Returns the ratio of number of types by number of tokens in a given 
string of words.
}
\examples{
get_variety(text=gardenParty)
}

