% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variety.R
\name{get_variety}
\alias{get_variety}
\title{Get type-token ratio/vocabulary richness measure}
\usage{
get_variety(text)
}
\arguments{
\item{text}{A vector of strings containing lines from a text.}
}
\description{
Returns the ratio of number of types by number of tokens in a given 
string.
}
\examples{
get_variety(text=gardenParty)
}

