% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_analysis_df.R
\name{make_analysis_df}
\alias{make_analysis_df}
\title{Create analysis table}
\usage{
make_analysis_df(text, chapters, freqwords, punctlist)
}
\arguments{
\item{text}{Character vector containing all the lines in a given text}

\item{chapters}{Lines to be matched in the text}

\item{freqwords}{Words whose frequency of appearance will be recorded}

\item{punctlist}{Punctuation characters whose frequency of appearance 
will be recorded}
}
\description{
This function takes a collection of chapters, breaks
them up into individual chapters, and returns summary
statistics for each chapter. Outputs average sentence length,
average token length, average type length, average variety 
(# type/# token), story length, number of occurences of words, 
number of occurances of punctuation and average sentiment score
(AFINN method).
}
\examples{
make_analysis_df(text="gardenParty",chapters=chapter.breaks, 
                 freqWords=c("the", "and", "a"))
}

