% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_analysis_df.R
\name{make_analysis_df}
\alias{make_analysis_df}
\title{Create analysis table}
\usage{
make_analysis_df(text, chapters, freqwords, punctlist)
}
\arguments{
\item{text}{Character vector containing all the lines in a given text}

\item{chapters}{Line index numbers associated with the first line/title 
line of each chapter}

\item{freqwords}{Words whose frequency of appearance will be recorded}

\item{punctlist}{Punctuation characters whose frequency of appearance 
will be recorded}
}
\description{
Takes a collection of chapters, breaks
them up into individual chapters, and returns summary
statistics for each chapter. Output statistics include average sentence length,
average token length, average type length, average variety 
(# type/# token), story length (number of words), number of occurences of words (expressed as
ratio of frequency of word over total number of words), 
number of occurances of punctuation (expressed as ratio of frequency of
punctuation mark over all punctuation marks) and average sentiment score
(AFINN method).
}
\examples{
breaks <- find_chapters(gardenParty)
words <- c("the", "and", "a")
punctuation <- c(",", ".", "...")
make_analysis_df(text = gardenParty ,chapters = breaks, 
                 freqwords = words, punctlist = punctuation)
}

