% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_length.R
\name{sentence_length}
\alias{sentence_length}
\title{Gets sentence length}
\usage{
sentence_length(sentences)
}
\arguments{
\item{sentences}{Vector of sentences}
}
\description{
Returns the length of a sentence in terms of number of
words.
}
\examples{
sentence_length(sentences=c("And after all the weather was ideal.", 
"Windless, warm, the sky without a cloud."))
}

