% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_length.R
\name{word_length}
\alias{word_length}
\title{Gets word length}
\usage{
word_length(words)
}
\arguments{
\item{words}{Vector of words}
}
\description{
Returns the length of a word in terms of number of
unique characters
}
\examples{
word_length(words=c("I", "am", "Sam"))
}

