% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_length.R
\name{word_length_line}
\alias{word_length_line}
\title{Gets word length per line}
\usage{
word_length_line(text)
}
\arguments{
\item{text}{Vector of strings representing lines of a text}
}
\description{
Returns a data table returning the average word length in a line
in terms of number of characters and the index number of that line.
}

